+++ tests/clients/conftest.py
@@ -1,12 +1,12 @@
 import pytest
 import requests_mock
 
-from trailblazer.clients.slurm_api_client.slurm_api_client import SlurmApiClient
+from trailblazer.services.slurm.api_service.slurm_api_client.slurm_api_client import SlurmAPIClient
 
 
 @pytest.fixture
-def client() -> SlurmApiClient:
-    return SlurmApiClient(
+def client() -> SlurmAPIClient:
+    return SlurmAPIClient(
         base_url="https://slurm.example.com",
         access_token="access_token",
         user_name="user_name",
