+++ tests/clients/test_slurm_rest_api_client.py
@@ -1,9 +1,9 @@
 from requests_mock import Mocker
-from trailblazer.clients.slurm_api_client.dto import SlurmJobsResponse
-from trailblazer.clients.slurm_api_client.slurm_api_client import SlurmApiClient
+from trailblazer.services.slurm.api_service.slurm_api_client.dto import SlurmJobsResponse
+from trailblazer.services.slurm.api_service.slurm_api_client.slurm_api_client import SlurmAPIClient
 
 
-def test_get_slurm_jobs(client: SlurmApiClient, mock_request: Mocker, jobs_response: dict):
+def test_get_slurm_jobs(client: SlurmAPIClient, mock_request: Mocker, jobs_response: dict):
     # GIVEN a mocked jobs response
     mock_request.get(f"{client.base_url}/slurmV0040/jobs", json=jobs_response)
