--- trailblazer/clients/slurm_api_client/dto/__init__.py
@@ -1,9 +1,0 @@
-from trailblazer.clients.slurm_api_client.dto.cancel_job_response import (
-    SlurmCancelJobResponse,
-)  # noqa: F401
-from trailblazer.clients.slurm_api_client.dto.job_response import (
-    SlurmJobResponse,
-)  # noqa: F401
-from trailblazer.clients.slurm_api_client.dto.jobs_response import (
-    SlurmJobsResponse,
-)  # noqa: F401
