+++ trailblazer/services/slurm/api_service/slurm_api_client/dto/__init__.py
@@ -1,0 +1,11 @@
+from trailblazer.services.slurm.api_service.slurm_api_client.dto.cancel_job_response import (
+    SlurmCancelJobResponse,
+)  # noqa: F401
+
+from trailblazer.services.slurm.api_service.slurm_api_client.dto.job_response import (
+    SlurmJobResponse,
+)  # noqa: F401
+
+from trailblazer.services.slurm.api_service.slurm_api_client.dto.jobs_response import (
+    SlurmJobsResponse,
+)  # noqa: F401
