+++ trailblazer/services/slurm/api_service/slurm_api_client/dto/cancel_job_response.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from trailblazer.clients.slurm_api_client.dto.common import SlurmError, SlurmWarning
+from trailblazer.services.slurm.api_service.slurm_api_client.dto.common import SlurmError, SlurmWarning
 
 
 class SlurmCancelJobResponse(BaseModel):
