+++ trailblazer/services/slurm/api_service/slurm_api_client/dto/common.py
