+++ trailblazer/services/slurm/api_service/slurm_api_client/dto/job_response.py
@@ -1,6 +1,10 @@
 from pydantic import BaseModel
 
-from trailblazer.clients.slurm_api_client.dto.common import SlurmError, SlurmJobInfo, SlurmWarning
+from trailblazer.services.slurm.api_service.slurm_api_client.dto.common import (
+    SlurmError,
+    SlurmJobInfo,
+    SlurmWarning,
+)
 
 
 class SlurmJobResponse(BaseModel):
