--- trailblazer/services/slurm/slurm_api_service.py
@@ -1,2 +1,0 @@
-class SlurmAPIService:
-    pass
